pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = ""
authors = ["Andrey Tarakanov <andrey.t-krd@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
gunicorn = {extras = ["eventlet"], version = "^21.2.0"}
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
requests = "^2.31.0"
flake8 = "^7.0.0"
flask-caching = "^2.1.0"
gevent = "^24.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
